+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+-   TODO: Describe initial release.
+
+## 0.0.2
+
+-   Incluindo compare(String e TextEditingController) em ValidatorUtil();
+-   Incluindo BaseState para usar em Base do StatefulWidget, onde o BaseState precisa <StatefulWidget, Provider>;